+++ b2sdk/raw_api.py
@@ -99,6 +99,7 @@ class AbstractRawApi(metaclass=ABCMeta):
         source_server_side_encryption: Optional[EncryptionSetting] = None,
         file_retention: Optional[FileRetentionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
+        cache_control: Optional[str] = None,
     ):
         pass
 
@@ -270,6 +271,7 @@ class AbstractRawApi(metaclass=ABCMeta):
         server_side_encryption: Optional[EncryptionSetting] = None,
         file_retention: Optional[FileRetentionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
+        cache_control: Optional[str] = None,
     ):
         pass
 
@@ -317,6 +319,7 @@ class AbstractRawApi(metaclass=ABCMeta):
         file_retention: Optional[FileRetentionSetting],
         legal_hold: Optional[LegalHold],
         custom_upload_timestamp: Optional[int] = None,
+        cache_control: Optional[str] = None,
     ) -> dict:
         headers = {
             'Authorization': upload_auth_token,
@@ -339,6 +342,9 @@ class AbstractRawApi(metaclass=ABCMeta):
         if file_retention is not None:
             file_retention.add_to_to_upload_headers(headers)
 
+        if cache_control is not None:
+            headers['X-Bz-Info-b2-cache-control'] = b2_url_encode(cache_control)
+
         if custom_upload_timestamp is not None:
             headers['X-Bz-Custom-Upload-Timestamp'] = str(custom_upload_timestamp)
 
@@ -359,6 +365,7 @@ class AbstractRawApi(metaclass=ABCMeta):
         file_retention: Optional[FileRetentionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
         custom_upload_timestamp: Optional[int] = None,
+        cache_control: Optional[str] = None,
     ):
         pass
 
@@ -708,6 +715,7 @@ class B2RawHTTPApi(AbstractRawApi):
         file_retention: Optional[FileRetentionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
         custom_upload_timestamp: Optional[int] = None,
+        cache_control: Optional[str] = None,
     ):
         kwargs = {}
         if server_side_encryption is not None:
@@ -724,6 +732,9 @@ class B2RawHTTPApi(AbstractRawApi):
 
         if custom_upload_timestamp is not None:
             kwargs['custom_upload_timestamp'] = custom_upload_timestamp
+
+        if cache_control is not None:
+            kwargs['cacheControl'] = cache_control
 
         return self._post_json(
             api_url,
@@ -891,6 +902,7 @@ class B2RawHTTPApi(AbstractRawApi):
         file_retention: Optional[FileRetentionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
         custom_upload_timestamp: Optional[int] = None,
+        cache_control: Optional[str] = None,
     ):
         """
         Upload one, small file to b2.
@@ -903,6 +915,11 @@ class B2RawHTTPApi(AbstractRawApi):
         :param content_sha1: hex SHA1 of the contents of the file
         :param file_infos: extra file info to upload
         :param data_stream: a file like object from which the contents of the file can be read
+        :param server_side_encryption: encryption setting for the file
+        :param file_retention: retention setting for the file
+        :param legal_hold: legal hold setting for the file
+        :param custom_upload_timestamp: custom upload timestamp for the file
+        :param cache_control: an optional cache control setting. Syntax based on the section 14.9 of RFC 2616. Example string value: 'public, max-age=86400, s-maxage=3600, no-transform'.
         :return:
         """
         # Raise UnusableFileName if the file_name doesn't meet the rules.
@@ -918,6 +935,7 @@ class B2RawHTTPApi(AbstractRawApi):
             file_retention=file_retention,
             legal_hold=legal_hold,
             custom_upload_timestamp=custom_upload_timestamp,
+            cache_control=cache_control,
         )
         return self.b2_http.post_content_return_json(upload_url, headers, data_stream)
 
@@ -960,6 +978,7 @@ class B2RawHTTPApi(AbstractRawApi):
         source_server_side_encryption: Optional[EncryptionSetting] = None,
         file_retention: Optional[FileRetentionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
+        cache_control: Optional[str] = None,
     ):
         kwargs = {}
         if bytes_range is not None:
@@ -1003,6 +1022,9 @@ class B2RawHTTPApi(AbstractRawApi):
 
         if file_retention is not None:
             kwargs['fileRetention'] = file_retention.serialize_to_json_for_request()
+
+        if cache_control is not None:
+            kwargs['cacheControl'] = cache_control
 
         try:
             return self._post_json(