+++ b2sdk/transfer/outbound/copy_manager.py
@@ -45,6 +45,7 @@ class CopyManager(TransferManager, ThreadPoolMixin):
         source_encryption: Optional[EncryptionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
         file_retention: Optional[FileRetentionSetting] = None,
+        cache_control: Optional[str] = None,
     ):
         # Run small copies in the same thread pool as large file copies,
         # so that they share resources during a sync.
@@ -60,6 +61,7 @@ class CopyManager(TransferManager, ThreadPoolMixin):
             source_encryption=source_encryption,
             legal_hold=legal_hold,
             file_retention=file_retention,
+            cache_control=cache_control,
         )
 
     def copy_part(
@@ -149,6 +151,7 @@ class CopyManager(TransferManager, ThreadPoolMixin):
         source_encryption: Optional[EncryptionSetting],
         legal_hold: Optional[LegalHold] = None,
         file_retention: Optional[FileRetentionSetting] = None,
+        cache_control: Optional[str] = None,
     ):
         with progress_listener:
             progress_listener.set_total_bytes(copy_source.get_content_length() or 0)
@@ -188,6 +191,7 @@ class CopyManager(TransferManager, ThreadPoolMixin):
                 source_server_side_encryption=source_encryption,
                 legal_hold=legal_hold,
                 file_retention=file_retention,
+                cache_control=cache_control,
             )
             file_version = self.services.api.file_version_factory.from_api_response(response)
             if progress_listener is not None: