+++ test/integration/test_raw_api.py
@@ -360,7 +360,8 @@ def raw_api_test_helper(raw_api, should_cleanup_old_buckets):
         {'color': 'blue'},
         io.BytesIO(file_contents),
         server_side_encryption=sse_b2_aes,
-        custom_upload_timestamp=12345,
+        #custom_upload_timestamp=12345,
+        cache_control='private, max-age=2222',
     )
 
     file_id = file_dict['fileId']
@@ -369,6 +370,9 @@ def raw_api_test_helper(raw_api, should_cleanup_old_buckets):
     print('b2_list_file_versions')
     list_versions_dict = raw_api.list_file_versions(api_url, account_auth_token, bucket_id)
     assert [file_name] == [f_dict['fileName'] for f_dict in list_versions_dict['files']]
+    assert ['private, max-age=2222'] == [
+        f_dict['fileInfo']['b2-cache-control'] for f_dict in list_versions_dict['files']
+    ]
 
     # b2_download_file_by_id with auth
     print('b2_download_file_by_id (auth)')