+++ test/unit/api/test_api.py
@@ -53,7 +53,9 @@ class TestApi:
     def test_get_file_info(self):
         self._authorize_account()
         bucket = self.api.create_bucket('bucket1', 'allPrivate')
-        created_file = bucket.upload_bytes(b'hello world', 'file')
+        created_file = bucket.upload_bytes(
+            b'hello world', 'file', cache_control="private, max-age=3600"
+        )
 
         result = self.api.get_file_info(created_file.id_)
 
@@ -84,7 +86,8 @@ class TestApi:
                 'serverSideEncryption': {
                     'mode': 'none'
                 },
-                'uploadTimestamp': 5000
+                'uploadTimestamp': 5000,
+                'cacheControl': 'private, max-age=3600',
             }
         else:
             assert isinstance(result, VFileVersion)