+++ test/unit/v0/test_bucket.py
@@ -283,7 +283,8 @@ class TestGetFileInfo(TestCaseWithBucket):
 
         self.assertIsInstance(info, FileVersionInfo)
         expected = (
-            a_id, 'a', 11, 'upload', 'b2/x-auto', 'none', NO_RETENTION_FILE_SETTING, LegalHold.UNSET
+            a_id, 'a', 11, 'upload', 'b2/x-auto', 'none', NO_RETENTION_FILE_SETTING,
+            LegalHold.UNSET, None
         )
         actual = (
             info.id_,
@@ -294,6 +295,7 @@ class TestGetFileInfo(TestCaseWithBucket):
             info.server_side_encryption.mode.value,
             info.file_retention,
             info.legal_hold,
+            info.cache_control,
         )
         self.assertEqual(expected, actual)
 
@@ -783,6 +785,21 @@ class TestUpload(TestCaseWithBucket):
         progress_listener = StubProgressListener()
         self.bucket.upload_bytes(data, 'file1', progress_listener=progress_listener)
         self.assertTrue(progress_listener.is_valid())
+
+    def test_upload_local_file_cache_control(self):
+        with TempDir() as d:
+            path = os.path.join(d, 'file1')
+            data = b'hello world'
+            write_file(path, data)
+            cache_control = 'max-age=3600'
+            file_info = self.bucket.upload_local_file(path, 'file1', cache_control=cache_control)
+            self.assertEqual(cache_control, file_info.cache_control)
+
+    def test_upload_bytes_cache_control(self):
+        data = b'hello world'
+        cache_control = 'max-age=3600'
+        file_info = self.bucket.upload_bytes(data, 'file1', cache_control=cache_control)
+        self.assertEqual(cache_control, file_info.cache_control)
 
     def test_upload_local_file(self):
         with TempDir() as d: